put all exercises unsolved.
