+++ buildspec.yml
@@ -16,8 +16,9 @@ phases:
       - echo "Installing dependencies..."
       - pip install -r tests/requirements.txt
       - apt-get install -y jq
-      - npm install -g allure-commandline --save-dev
-      - allure --version
+#      TODO Exercise 3: Install allure:
+#      - npm install -g allure-commandline --save-dev
+#      - allure --version
 
   pre_build:
     commands:
@@ -33,30 +34,31 @@ phases:
     commands:
       - echo "Running pytest tests..."
       - env
-      # - pytest tests 
+      - pytest tests
       # Exercise 3: Add allure report generation and upload to S3
       # comment previous line and uncomment the following line to run allure report generation and upload to S3
-      - pytest tests --alluredir=$ALLURE_RESULTS_DIR 
-      
-      - echo "Running Great Expectations tests..."
-      - python tests/great_expectations/raw_validation/main.py  # Execute the Great Expectations tests
-      - python tests/great_expectations/clean_validation/main.py  # Execute the Great Expectations tests
+#      - pytest tests --alluredir=$ALLURE_RESULTS_DIR
+
+#      # Exercise 4: Execute Great Expectations tests
+#      - echo "Running Great Expectations tests..."
+#      - python tests/great_expectations/raw_validation/main.py  # Execute the Great Expectations tests
+#      - python tests/great_expectations/clean_validation/main.py  # Execute the Great Expectations tests
       
 
   post_build:
     commands:
     # Exercise 3: Add allure report generation and upload to S3
-    # uncomment the following line to run allure report generation and upload to S3
-      - echo "Ensuring S3 bucket structure exists"
-      - aws s3api head-object --bucket $ALLURE_REPORT_BUCKET --key allure-report/ || aws s3api put-object --bucket $ALLURE_REPORT_BUCKET --key allure-report/ 
-      - echo "Downloading existing Allure history"
-      - aws s3 sync s3://$ALLURE_REPORT_BUCKET/history/ $ALLURE_RESULTS_DIR/history/ || true
-      - echo "Generating Allure report"
-      - allure generate $ALLURE_RESULTS_DIR -o allure-report --clean
-      - echo "Uploading Allure report to S3"
-      - aws s3 sync allure-report s3://$ALLURE_REPORT_BUCKET/ --delete
-      - echo "Updating Allure history"
-      - aws s3 sync allure-report/history s3://$ALLURE_REPORT_BUCKET/history/ --delete
+    # uncomment the following lines to run allure report generation and upload to S3
+#      - echo "Ensuring S3 bucket structure exists"
+#      - aws s3api head-object --bucket $ALLURE_REPORT_BUCKET --key allure-report/ || aws s3api put-object --bucket $ALLURE_REPORT_BUCKET --key allure-report/
+#      - echo "Downloading existing Allure history"
+#      - aws s3 sync s3://$ALLURE_REPORT_BUCKET/history/ $ALLURE_RESULTS_DIR/history/ || true
+#      - echo "Generating Allure report"
+#      - allure generate $ALLURE_RESULTS_DIR -o allure-report --clean
+#      - echo "Uploading Allure report to S3"
+#      - aws s3 sync allure-report s3://$ALLURE_REPORT_BUCKET/ --delete
+#      - echo "Updating Allure history"
+#      - aws s3 sync allure-report/history s3://$ALLURE_REPORT_BUCKET/history/ --delete
 
 # artifacts:
 #   files:
