--- requirements.txt
@@ -1,3 +1,0 @@
-pytest==8.3.2
-boto3==1.34.150
-great_expectations==0.18.19
